pyproject.toml
[project]
name = "microblog"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Виктор Сагалаков", email = "sagalakov.viktor@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "alembic>=1.16.5",
    "uvicorn>=0.34.2",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.21.0",
    "python-multipart>=0.0.20",

]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
